mountinfo: add GetMountsFromReader
Some code have tests or benchmarks that come with fake mountinfo.
To switch those to use moby/sys/mountinfo without any major rework
of those tests and benchmarks we need to provide a way to work
with such fake mountinfo.

This is probably more important for benchmarks, since without
this we can't make any comparisons after switching to moby/sys.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>