mountinfo: fix path unescaping
Moby PR #38977 [1] and containerd PR #3159 [2] added the handing of
escape sequences in mountinfo paths (root and mountpoint fields).

Unfortunately, it also broke the handling of paths containing double
quotes, as it was pointed out in [3].

The solution is to stop using strconv.Unquote and write our own
specialized function to deal with escape sequences.

Unit tests added.

[1] https://github.com/moby/moby/pull/38977
[2] https://github.com/containerd/containerd/pull/3159
[3] https://github.com/containerd/containerd/issues/4257

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>