mount.RecursiveUnmount: add a fast path
In many cases, this is called with an argument that is a mount point
itself. In such case, if we're on Linux, MNT_DETACH ensures that the
umount is recursive (which is its nice, but a little known feature).
For other platforms, if the mount does not have any submounts
underneath and is not busy, it will be unmounted.

In the above case, we no longer have to parse mountinfo and perform
the unmounts one-by-one.

IOW this should speed up RecursiveUnmount() for some cases, without
breaking the compatibility.

While at it
 - fix a typo in doc;
 - add that the argument does not have to be a mount point.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>