mount.Mounted: normalize the path argument
Currently, the path provided to mount.Mounted() must be
 - an absolute path;
 - with all symlinks resolved;
 - cleaned.

Otherwise, the function will most probably return a false negative in
case of a bind mount (i.e. "not mounted" while the path is in fact
mounted).

This commit removes the above limitation by doing a path normalization.
This should slow things down but it's not affecting the fast path, so
the slowdown will only be seen in the bind mount case.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>