mountinfo.Mounted: optimize on Linux with openat2
A new (as of Linux 5.6) openat2(2) system call can be used to check
whether a path is a mount point or not, and it even works for bind
mounts!

This relies on the fact that openat2() with RESOLVE_NO_XDEV flag
won't cross the filesystem boundary (that includes bind mounts).

If EXDEV error is received, it means the argument is definitely a mount.
If no error is received, it means the argument is definitely not a
mount. For all the other errors (most probably ENOSYS or EPERM), fall
back to existing methods (mountedByStat,  mountedByMountinfo).

Add a couple of tests:

* TestMountedBy prepares a few mounts, and checks that various
  mountedBy* functions work as expected. Requires root.

* TestMountedByOpenat2VsMountinfo reads the mountinfo and then tries
  to use mountedByOpenat2() for each mount, ignoring EPERM as the test
  might not be run as root. AFAIK this is the only way to test the
  new functionality if one is not root.

v2 changes:
 - drop our own openat2 wrapper;
 - add O_PATH to openat2 flags;
 - add O_CLOEXEC to all open* calls;
 - rm socket and symlink special cases.

v3:
 - drop the stat parameter, remove the dir/file distinction;
 - use openat2 for dirfd as well, so we can bail out earlier.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>