mountinfo.mountedByMountinfo: normalize path
Currently, the path provided to mountinfo.Mounted() must be
 - an absolute path;
 - with all symlinks resolved;
 - cleaned.

Otherwise, the function may return a false negative (in case this is
a bind mount and the kernel is too old to have openat2) -- i.e. it
will return "not mounted" while the path is in fact mounted.

This commit removes the above limitation by doing a path normalization.
This should slow things down but it's not affecting any of the fast
paths, so the slowdown will only be seen in the "bind mount and older
kernel" case (which is already pretty slow since we have to parse
mountinfo).

v2: as in mountedByOpenat2 and mountedByStat, treat ENOENT as "not
mounted".

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>