mountinfo: treat ENOENT as "not mounted"
The positive side effect from this is we're returning earlier, not
falling back on a slow path.

The downside is this might not be entirely true, as a mounted path can
be overshadowed by another mount, but still present in mountinfo.

As the mount is inaccessible anyway, there is nothing that can be done
with it, so let's treat it as non-mount.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>