mountinfo: use idiomatic naming for fields
I am looking at using these modules in containerd, so that we can
reduce the effort of maintaining "mountinfo" implementations.

This change renames the info fields to match the ones used in
containerd, which are more "correct" from a Go naming perspective:

https://github.com/containerd/containerd/commit/8cd218237b49b41bffa0035393143571ec01b157

This will be a breaking changes for consumers of this package, but
we're still pre-v1.0.0, so if we want to make these changes, this
is still possible now.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>