Make GetMountsFromReader linux-only
GetMountsFromReader was a wrapper for parseInfoFile on Linux.
All other implementations do not parse `mountinfo`, so had to
stub out the implemetation and omit the actual reader.

(worse, a FreeBSD implementation was missing, causing compile on
that to fail).

This patch makes GetMountsFromReader linux-only, as it was added
to allow benchmarking the `mountinfo` parsing, which only makes
sense on Linux.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>