mountinfo: deprecate PidMountInfo
Function PidMountInfo is not named correctly (it should be PID not Pid,
and we already have mountinfo in the package name so it's tautological),
and can not accept FilterFunc.

Besides, there are not too many users of this function (I was only able
to found one, see [1]) and users can switch to GetMountsFromReader.

Depreate this function. The plan is to remove it before v1.0.

[1] https://github.com/DataDog/datadog-agent/pull/6495/commits/406d98801434dedc87

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>