Create proper mounter for OpenBSD
OpenBSD's mount(2) works differently since it uses separate structs to
present options for different filesystems, unlike in FreeBSD. It also
means that support for other filesystems need to be added explicitly.

I'm not sure which ones are required for now so I just write support
for only FFS, which is the go-to filesystem used by OpenBSD systems.

Signed-off-by: Muhammad Kaisar Arkhan <hi@yukiisbo.red>