mountinfo: GetMountsFromReader() remove workaround for kernel < 3.10
In Linux <= 3.9 mounting a cifs with spaces in a share name (like "//srv/My Docs")
_may_ end up having a space in the last field of mountinfo (like "unc=//serv/My Docs").

Since kernel 3.10-rc1, cifs option "unc=" is ignored, so spaces should not appear.

Given that all major distros are now on at least kernel 3.10, and (e.g.) docker
has not supported older kernel versions for 5 Years [1], it should be save to
remove this special handling.

[1]: https://github.com/moby/moby/commit/51b23d88422918c24291f5876df35f91b23a446a

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>