mountinfo: unify mountedBy checks
Instead of copy/pasting the code to check individual mountedBy*
implementations, add a list and a loop to check them.

The only special thing is, we have to check if the name of the function
being tested is "mountedByStat", to skip erroring out on bind mounts.
Otherwise the code is the same.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>