mountinfo: add MountedFast
MountedFast is a method of detecting a mount point without reading
mountinfo from procfs. A caller can only trust the result if no error
and sure == true are returned. Otherwise, other methods (e.g. parsing
/proc/mounts) have to be used. If unsure, use Mounted instead (which
uses MountedFast, but falls back to parsing mountinfo if needed).

If a non-existent path is specified, an appropriate error (rather than
"not mounted") is returned. In case the caller is not interested in this
particular error, it should be handled separately using e.g.
errors.Is(err, os.ErrNotExist).

This function is only available on Linux. The implementation mostly
relies on openat2(2) syscall, available since Linux kernel v5.6.
On older kernels, this uses stat(2) syscall which can reliably
detect normal (but not bind) mounts.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>