Makefile: rm .SHELLFLAGS, add set -e
Apparently GHA CI Mac OS X environment uses old GNU Make (v3.81) which
does not yes support .SHELLFLAGS (added by commit 8589e2e7).

Since we call a few commands (under "for") from the make target,
intermediate failures are ignored unless "-e" shell option is set
(or explicit "&&" is used, which is not possible to do in "for").

Example:

	for cmd in false true; do $cmd; done

This code succeeds (has exit code of 0), unless "set -e" is set, because
the last command executed is "true" with the exit code of 0.

Use explicit "set -e" where needed. While at it, also add "-u" to catch
possible uses of uninitialized variables.

Also, replace && with ; for multiple commands. Technically this is not
required for the fix, but since having set -e results in implicit &&,
there is no need to have && explicitly. In other words, this change is
more like a reminder that we do not have to use && here.

Fixes: 8589e2e7
Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>