capability: Remove init and CAP_LAST_CAP, add LastCap
One (minor) issue with this package is it has func init which reads a
file in /proc, making the start of any program which imports the package
a bit slower.

Let's switch to lazy initialization, i.e. only read the file when needed.
Unfortunately, this can not be done in a non-disruptive manner, since we
have a public variable CAP_LAST_CAP.

So, this is a disruptive change, and anyone who's using CAP_LAST_CAP
should change their code to something like this:

	last, err := capability.LastCap()
	if err != nil {
		return err
	}

Also, add a test case for LastCap.

Reported-by: ningmingxiao <ning.mingxiao@zte.com.cn>
Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>