capability: deprecate List, add ListKnown, ListSupported
Apparently, most users of capability.List wants the list of supported
capabilities (i.e. they go on to exclude capabilities above the last
known one).

Let's provide ListSupported to such users.

Also, provide ListKnown, and deprecate List.

Finally, amend LastCap documentation with a link to ListSupported,
as this is what most users are using it for.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>