capability: test nits
First, make capability_test.go a separate package so we are limited to
testing the external API (as opposed to e.g. state_linux_test.go in
which we test internal stuff as well).

Second, streamline the linux/non-linux code so it's easier to read and
copy/paste (in general, we only want to check that non-linux functions
can be called and return an error).

Third, change some non-fatal errors to use t.Errorf.

Fourth, use want/got everywhere in error messages, and add more context
to some errors.

Nothing here is critical or makes any noticeable impact on tests.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>