ci: don't install private golangci-lint
First, let's not install golangci-lint from Makefile, instead relying on
whatever version a developer already has installed and available.

Second, we still need golangci-lint in GHA CI. For that, let's use
golangci-lint-action. Unfortunately, there's no way to NOT run it
from the action, or run it on multiple directories (see [1]), so
run it on a single module.

This is better for CI because we use all the good bits from the
golangci-lint-action (caching and annotations).

While at it, bump golangci-lint from v2.0.3 to v2.4.x.

[1]: https://github.com/golangci/golangci-lint-action/issues/1226
Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>