uname: add
This is a tiny package to check the Linux kernel version. It is taken
from the golang sources (src/internal/syscall/unix), with the following
changes:

 - removed irrelevant parts, including FreeBSD and Solaris kernelVersion
   implementations;
 - added a tiny test case for KernelVersion;
 - added some documentation;
 - added proper copyright headers.

The code is covered by the "BSD 3-clause" license, which is compatible
with Apache license.

The history of the code before the fork can be seen here (oldest first):
 - https://go-review.googlesource.com/c/go/+/424896
 - https://go-review.googlesource.com/c/go/+/427675
 - https://go-review.googlesource.com/c/go/+/427676
 - https://go-review.googlesource.com/c/go/+/700796

This is aimed to replace the relevant containerd package
(github.com/containerd/containerd/pkg/kernelversion) and its forks.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>